test(core): Cleaner Document tests
